requirements.txt
opencv-python
mediapipe
pyautogui
pynput
screen-brightness-control
pycaw
comtypes
Pillow
numpy
